backend/requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
google-generativeai==0.3.0
python-multipart==0.0.6
pydantic==1.10.10
pytest==7.4.0
httpx==0.24.1
pytz==2023.3
black==23.3.0
flake8==6.0.0
PyPDF2==3.0.1